pyproject.toml
[project]
name = "discobot-aggt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
